[HW] Update new builder to not call `erase` on OutputOp directly.
This builder might be used with a PatternRewriter, which requires all
IR mutation such as erases to go through the PatternRewriter. Instead,
add a flag to the generic builder to decide if an OutputOp terminator
should be created. In the new builder, this is set to false, and the
correct OutputOp is created directly, rather than erasing the old one
and creating a new one.