[PrepareForEmission] Reorder inout ops before legalizing OoO uses (#3865)
This commit fixes an assertion failure in PrepareForEmission that wrongly
tries creating temporary wires for inout operations by moving an inout operation 
after its operand position before actually resolving out-of-order uses.

In the first place, inout operations are *always inline operations* therefore we might not have to
legailze OoO uses of inout operations. But it would be more reasonable to keep the
invariance that "after PrepareForEmission, operations in the toplevel module are
sorted in the def-use order".