[PrepareForEmission] Fix "always inline" operation legalization
This fixes a bug that "always inline" operation is not legalized for the
last user. Previously `lowerAlwaysInlineOperation` skipped moving the op
for the last user but it caused a bug. If the last user is placed before
the op, it causes an error in out-of-order resolution. This fixes the
bug by removing the condition `lowerAlwaysInlineOperation` and avoiding
processing same operations twice to prevent infinte loops.