[FIRRTL] Lower Memory debug port to RefType of Register (#3911)
This commit updates the memory lowering passes to handle the memory debug port.
The `MemToRegOfVec` pass lowers the debug ports to a RefType of the register.
The pass connects the debug ports to a `RefSend` of the register. This ensures
 that any `MemTap` of the memory can be lowered to a cross-module-reference of
 the memory register.
The `FlattenMemory` cannot flatten a memory with any Debug port.
The `LowerTypes` cannot handle debug port in memories with aggregate data type.