Convert binary muxes to ifs when feeding seq.firreg (#3925)
Traditionally, firrtl lowered muxes feeding registers to if statements.  Now that the firrtl spec is updated to indicate firrtl is two state, muxes are lowered to comb 2-state muxes, which lets us lower those muxes to if statements.  This is required for some power tools to pattern match clock gates.