[PrepareForEmission] Fix logic op position legalization
This PR fixes an issue of logic operation legalizations. Previously
an iterator is managed while executing legalizations, but it is very
difficult to properly manage the iterator since a lot of operations are
inserted into various places. To fix the issue, this separates the legalization
into another loop which is executed only when in a procedural region
to move logic operations to valid points.