[HW] Enable `i0` for HWIntegerType
Luckily a fairly minor change due to ExportVerilog already have good support for emitting i0-valued expressions as comments.
This commit ensures that `hw.array_get` for singleton arrays may be emitted legally, which is probably the most pressing matter in embracing `i0`-typed values.
Not a solution to every problem in the code-base related to this, but I think this is a reasonable starting point.

Changes:
- zero-width `mlir::IntegerType` valid as HWIntegerType's
- `hw.constant` generates `i0` constants if provided an `i0` type (no change)
- `hw.array_get` now expects `i0` type for singleton arrays.
- `hw.array_get` will always emit `1'b0` for `i0`-typed indexes.
  The expression corresponding to the `i0` index is emitted inlin as a comment.
- Some changes to FIRRTL subindex and subaccess lowerings