[HandshakeToHW] Add load,store conversion + dummy ext memory lowering
External memories are... tricky... to handle in the pattern rewrite infrastructure due to the need for rewriting the single top-level memref argument to both new in- and output ports of the module (HW doesn't contain flip types, like FIRRTL).

That will, therefore, require a more specialized lowering of external memories. As an intermediate step, load and store ops can be elaborated, the external memory emitted as an external module, and the top-level `memref` argument as an `esi.channel<memref>`.

A follow-up commit will then implement lowering of the `extmemory` + top-level `memref` by elaborating the ports required for that `memref` and performing the required plumbing.