[PrepareForEmission] Perform wire spilling based estimated expression size (#3752)
This PR implements mechanism to perform extra wire spilling after legalization in Prepare. 
After legalization, we are able to know (almost) accurate verilog AST structures, for example we can 
calculate accurate sizes of verilog expressions. Obviously we want to use this kind of information to emit 
more readable&debuggable outputs. As a first step, this PR adds `wireSpillingHeuristic` options to 
specify style of extra wire spilling. There is a trade off between simulation performance and readablility so 
it would be good for users to choice the output style. Currently only `spillLargeTermsWithNamehints` option is 
implemented. The option spills wires for expressions with namehints if their sizes are greater than the given threshold.