[circt-lec] Introduce the `circt-lec` tool
A logical equivalence checking tool for CIRCT.
As of now it covers some basic `hw` operations and the whole `comb` dialect.
Z3 acts as the logical backend and has been introduced as a build dependency
for the tool. Furthermore, regression tests for part of the implemented
operations and features have been written.