[HW] Also modify HWModule block args if adding/removing inputs (#3997)
Don't know if this is a feature or a bug - `modifyModulePorts` currently only modifies the top-level I/O of a module, but does not change the block arguments of the module body. This leaves performing module body block arguments up to the caller, which is a fairly cumbersome task given that all of the bookkeeping necessary to do this is kept within `modifyModulePorts`.
AFAICT this hasn't been since no codepath in CIRCT currently leverages this.
Fixed by adding/removing block arguments if a `Block*` is provided to `modifyModulePorts`.