[FIRRTL] Fix GCT Instance Name Prefix (#4016)
Stop mangling the name of Grand Central (GCT) Views in the companion and
parent annotations during the PrefixModules pass.  This fixes a bug
where the prefix was incorrectly applied to the instance name of an
interface.

Extend an existing end-to-end GCT View test case to also check that
NestedModulePrefixAnnotation and PrefixInterfacesAnnotation apply and
compose correctly.  Specifically, the prefix annotation is applied to
all module names, but the prefix interfaces annotation is only applied
to the interfaces.  The instance name should match the name of the view
and not be prefixed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>