[ExportVerilog] Fix use of wrong decl alignment after emitting a block.
Before emitting a block of statements, we scan for all declarations
(wires, regs, ...) and calculate the max width of the types and names
so we can align the types and names neatly across declarations.

Previously declarations were (generally) gathered and emitted at the top,
but now occur throughout (new emission mode).

There's a bug with how this works presently:

When emitting nested blocks of statements (always, etc.) the newly
calculated widths override the existing ones-- resulting in
wires/regs after the block being aligned using the widths from
declarations within the block-- instead of using widths calculated
for, and from, the declarations outside the block.

To fix this, simply save the widths and restore them after the block.

These declarations are sometimes a bit far apart for the alignment
to always be obvious locally, but as long as we align them all
together per-block let's use the right alignment values not those
from any nested blocks emitted along the way.