[Seq] Add FirReg no-update canonicalizer
When a FIRRTL register has no next value, we optimize it away by
replacing it with its reset value if it has one, or zero.  This adds a
`seq.firreg` canonicalizer to cover this situation.  This helps fix a
phase ordering problem between the more powerful canonicalizers that
exist in the `comb` dialect and where this optimization is applied in
`firrtl` dialect canonicalizers.