[FIRParser] Stop double deserializing OMIR annos (#4062)
When we import OMIR annotations and turn them into MLIR attributes, we
are accidentally doing twice it back to back. This is some dead code
that was left behind in the transition to the `LowerAnnotationPass`.
Removing this code does not require any test changes. On a large core
with a lot of OMIR this improved the OMIR parsing speed by 15%.