[LowerToHW] Lower subindex into array_get instead of read(array_index)
This PR changes the way of lowering of subindex. Currently a
subindex operation is lowered into an array index inout operation.
When these subindex are used as source values, we dereference the
inout operations with read_inout. Alternatively this PR lowers a
source value subindex into an array_get operation. This improves the
quality of output verilog since usually dataflow analysis is blocked by
inout operations.