[ExtractTestCode] Support extracting instances in some cases. (#3522)
This extends the set of what is extracted to include instances. When
it can be shown that the fan-out of an instance is only used in the
extracted slice, the instance and its fan-out are removed, rather than
left behind.

This allows some cases of instances that only existed to feed into the
test code to now be removed, which prevents leaving behind dead code
that may trip lint warnings.