[SV] Update ExtractTestCode to inline input only modules. (#4063)
PR #4014 updated ExtractTestCode to find input only modules, move them
to the testbench, and bind their instances in. This isn't actually
allowed by the System Verilog spec, which prohibits nested bindings.

Instead, this PR inlines instances of input only modules when
possible. If all instances of such a module are inlined, the module is
removed.