test: tweak test discovery to enable unittests, fix deps var.
Discovery will find the generated lit.site.cfg.py, which
knows how to load the lit.cfg.py.

This could be gated on whether we're building "standalone",
as flang does, but that still doesn't work as we are not part
of the config mapping.

cc #2750.

Also fixup typo in variable name, .._DEPS -> .._DEPENDS.