[SV] Fix heap-use-after-free in recent SVExtractTestCode change.
We iterate over the instances of a module in the instance graph, and
during that iteration, erase some of the uses. This causes a
heap-use-after-free that ASAN detects. A simple fix is to use
llvm::make_early_inc_range, which allows us to safely iterate over a
range while we erase things from the range.

Fixes https://github.com/llvm/circt/issues/4081.