[SV] Fix potential to erase ops multiple times in SVExtractTestCode.
When we decide to extract an instance, we must erase it and its
forward dataflow. However, it's possible for the forward dataflow to
contain nested ops. In this case, we might try to erase children after
they've been erased as part of their parent. The erasing loop is
updated to track all ops that are already erased to avoid this.

There is also an issue with having the erasing logic buried inside
doModule, because the new code to conditionally inline some modules
might try to clone ops that have been erased. Instead, the set of ops
to erase is threaded through all calls to doModule as well as
inlineInputOnly, and only after all the IR mutation is complete, do we
finally clean up by erasing the instances and their forward dataflow.