[SV] Fix SVExtractTestCode to handle inlining cycles.
This was raised on the issue that originally landed the inlining
feature, but we weren't able to find an example of this in the
wild. Of course, it has turned up: some modules only have inputs, feed
test code, and thus meet the criteria for inlining. But they also
might have internal signals that are not extracted, and form cycles in
the SSA graph. The test case is the most trivial example, but more
complex structures also hit the same bug and are fixed in the same
way.

The solution is to use the same helpers that the main extraction logic
uses to deal with comb cycles. These can be dropped in to the little
inlining helper as-is and resolve the issue.