[HW] Refactor InstanceOp implementation to enable code reuse
Move implementation code of the InstanceOp to a new
InstanceImplementation file such that it can be included and reused in
other dialects that need to implement a similar instance operation.
It tries to extract code parts in a way that basic checks like same
length of operand and argNames arrays, matching types, etc. are
performed in a way that it does not need to know about the module or
instance operations and thus are also applicable in more generic cases.
However, it also provides functions that require the instantiated module
to be a HW module and thus allow for more code reuse for instance
operations that can only refer to those.
Useful custom directives for table-gen assembly formats are collected in
a separate file that should only be included in cpp files, while the
InstanceImplementation could also be included in the header file.