Link gtest against portable threading lib
When we build CIRCT as a standalone project (not a unified build), we
build gtest from scratch for our unit tests.  This is currently
hard-coded to link against pthreads, which does not work on Windows. We
can use the built in `Threads::Threads` cmake library which is more
portable.  This fixes the unit test linking on Windows.