[SV] Fix SVExtractTestCode to extract all results of instances.
When we decide we can extract an instance, we must remove results of
the instance from  the set of inputs to the extracted instance.
Previously, we only removed one result at a time, as we iterated over
the set of inputs. This is problematic when instances have multiple
results, and all of them feed the test code. Because we add the
instance to the test code set the first time we encounter one result
and decide the instance should be extracted, the next iterations that
visit the same instance do nothing, since we've already marked it for
extraction.

The solution is simple: as soon as we've proved we can extract an
instance, remove all of its results from the set of inputs to the
extracted instance.