[SV] Fix SVExtractTestCode to not duplicate when inlining. (#4106)
We previously switched from eagerly erasing ops to clean up to lazily
doing so after all IR mutations to fix a bug. This is important for
any instances we extract and their forward dataflow, but as part of
that change we also made the erasure of test code roots
themselves lazy (the assert, assume, and cover statements). This
causes issues in case any module with test code gets inlined later;
the test code roots are revisited, extracted again, and
duplicated. This is illegal, because some test code may include side
effects, like printing to the console.

The solution is simple: make sure to eagerly erase all test code
roots, and remove them from the set of ops to erase lazily, if they
were reachable from an instance that is being extracted. The latter
change is to prevent doubly erasing the same op.