[FIRRTL] Add ResolveTraces Pass
Add the ResolveTraces pass that implements Chisel's TraceName API.  This
pass updates TraceName/TraceAnnotations with a "target" field indicating
the current FIRRTL Target value of the thing the Annotation is attached
to.  The pass then writes these Annotations to a JSON verbatim blob
outside the circuit that will be written to a file derived from the
Circuit name.  The original Annotations are removed.

This API works by enabling users to attach "traces" to their Chisel
design to enable them to recover the final target in the compiled
Verilog.  This is intended to the be the low-level basis on top of which
Chsiel users can build their own tools that require information about
the final locations of things in Verilog, e.g., generating TCL scripts.

Necessarily add both TraceNameAnnotation and TraceAnnotation, part of
Chisel's Trace API, to AnnotationDetails and enable the LowerAnnotations
to resolve and apply them.

TraceNameAnnotation is a deprecated API that can be used to track a
Chisel-named thing with don't touch behavior.  TraceAnnotation is an API
that canb e used to track a Chisel-named thing and does NOT have don't
touch behavior.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>