[SV] Fix SVExtractTestCode to not duplicate instances when inlining.
If the instance has been extracted, it shouldn't also be cloned when
inlining. Since inlining runs after all extraction, during inlining,
simply check if any instance or its forward dataflow is in the set of
ops to be erased, and mark the inlined op to be erased as well. We
have to copy over ops we know are dead to ensure the IR remains valid
until it is cleaned up.