[SV] Fix SVExtractTestCode to keep the instance graph up to date.
Notably, when inlining the body of a module, any instances in the body
should be added to the new parent module in the instance graph. The
instances in the old body are already being removed from the instance
graph when we remove the old module. This fixes an issue where
instances that were inlined appeared to have no uses, and were
erroneously moved to the testbench.

This also makes sure to add modules created to hold extracted test
code to the instance graph. This bookeeping is important so the above
can work, and find the newly created module during inlining.