[ExportVerilog] Don't combine location info across if/elseif chains.
When emitting IfOp, we recursively look for another IfOp in the else
branch so we can emit a nicer `else if`.

While doing this, reset the set of operations contributing to location
information to those for the active IfOp and its condition.

The final `else` will have the information for the last IfOp processed.