[Seq] Add behavioral lowering of `seq.hlmem`
This commit adds a fallback/default pattern for lowering `seq.hlmem` operations. The implementation is based on the ESI ram service lowering, and provides a behavioral lowering. The intention of this lowering is to provide a pattern that will almost guarantee to synthesize on FPGAs - QoR not guaranteed. "Better"/target-specific patterns can be added by adding patterns with higher benefit - which will match with higher priority.

Read ports with arbitrary (N) latency are realized by performing a combinational read, and registering the read value N times.
Write ports are only supported for latency == 1.

tidy