[ExportVerilog] For parameter values, handle parens separately.
Instead of combining "subprecedence" with check for whether
this expression needs parentheses, split these up and check
the right one for whether to add outer parentheses.

Operators have two precedence levels of concern:
* subprecedence, specifies when parens needed around parens.
* prec, for determining if this expr needs parens around it.

For binary and unary operators the two precedence levels are the same.

Combine function operators (clog2) and (Str)Concat into "hasOpenClose" for
operators that add syntax around their operands.

Set subprecedence appropriately for (Str)Concat, although as long as used
for strings this doesn't change anything.

Not sure of the outer precedence for $clog2, so set to Symbol.