[FIRRTL] Fix GCT output dir for mod/extmod
Change Grand Central (GCT) View behavior that determines which output
directories are used for blackboxes which are instantiated by both
inside and outside the View's companion module.  Previously, if in both
then the blackbox was placed in the View's directory.  This is changed
to not modify the output directory of the blackbox.  Only blackboxes
which are _exclusively_ instantiated by the View are moved.

Change modules to also have the above behavior.  Previously, module
submodules were not moved.

This is implemented by using graph iterators thanks to some great ideas
from @youngar.

This moves closer to SFC behavior.  In the SFC, modules instantiated
exclusively in the View are moved.  Modules instantiated in both are
copied in both output directories.

Co-authored-by: Andrew Young <youngar17@gmail.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>