[FIRRTL] Only Set Directories in GCT Views
Change Grand Central (GCT) Views pass to only add directory information
to companions, blackboxes, and interfaces.  This enables the final
filename to be set by ExportVerilog which has a better idea of what
things should be named and guarantees that the final module name is
going to match the filename.

h/t @youngar

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>