[SV] Implement a canonicalizer to merge neighboring procedural assignments (#4074)
This PR implements a canonicalizer to merge neighboring procedural assignments, e.g:
```
a[0] <= b[1]
a[1] <= b[2]
a[2] <= b[3]

->
a[2:0] <= b[3:1]
```

For a procedural assignment, we first check the next operation is the same kind of an assignment. If so, we check both of dest values are array indexing inout operations that point at neighboring elements.