[FIRRTL] Significantly improve performance setting port syms. (#4223)
Don't create many attributes (even empty ones) just to throw
them away.  Also, cache the single entry if creating any,
save cost of repeatedly uniquing in `Attribute::get`.
This is significant when there are thousands of ports
(and if this is done thousands of times).

On one design, improves speed of LowerAnnotations 720s -> 640s.

There's still some copying/array-fill going on that could be tightened,
(especially in case where array is all valid or empty, and/or the caller
already has a suitable `SmallVector<>`.)
but this gets this code from appearing constantly in sampled stacks.

Tweak `setPortSymbolAttr` to use `fixupPortSymsArray` and set the
attribute directly to avoid copying the array it just created.

Also check for empty (but non-NULL) InnerSymAttr.