[FIRRTL] Guarantee NodeOp in GCT View
Change Grand Central (GCT) Views Annotation application logic to always
create a dummy node to store the value of something that is a field of
an interface.  This makes logic during the actual GCT pass simpler (and
easier to check) as this now guarantees that the pass will always see a
node with a DontTouchAnnotation, even for designs where the view is
composed of things inside the companion module.

Relax an overly aggressive check in the Grand Central (GCT) Views pass
which was incorrectly asserting that the source of a SystemVerilog
Interface's field had to be a node of a RefResolveOp or ConstantOp.
This can actually be anything and it will still work.

Fixes #4277.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>