[FIRRTL][Dedup] Fix cloning annotations for more context (#4285)
When we merge two modules, sometimes we have to add more context to a
non-local annotation.  When this happens, we create a new hierarchical
path operation and update all annotations to use the new path.  In order
for us to find all annotations which need to be updated, we store a
mapping, `targetMap`, from each HierPathOp to the annotations that
reference them.  This fixes a bug where we were not keeping the
`targetMap` updated when we copy a non-local annotation when deduping
operations. This is a regression from
46e9b214329de820368ecf2956bd0c3aff43cf36, where we stopped updating the
target map.  See
https://github.com/llvm/circt/commit/46e9b214329de820368ecf2956bd0c3aff43cf36#diff-d0c16a1337e2a6200ae0f4c763fadb7ddaa3f8b389130aa04cd1e61ff2fba7eaL896
for the exact line that was removed.