[SV] Update SVExtractTestCode to handle already bound instances.
We already skip over extracting from already bound instances to
prevent binds under binds. And we already do not inline or extract
instances that have an inner ref. But sometimes we extract test code
from a module that contains a bound instance sibling to the newly
bound test code instances, and then inline an instance of that
module.

In that case, the bind needs to be updated. To support this, the
BindTable is pre-populated with already bound instances, and the
existing logic to update binds during inlining fixes them up.