[IMDCE][SV] SV-level inter-module dead-code elimination
This pass implements a basic inter-module dead-code elimination pass.
Based on a liveness analysis, ops which can be deleted and have no
anchored references (not used by public ports or ops what cannot be erased)
are eliminated from the design. The port lists of modules are also adjusted
accordingly.

Some more feedback and iteration is required to correctly pin down all ops
which we do not want to remove.