[FIRRTL] Change Wires to Nodes in WiringProblems (#4312)
Change the connect utility lambda used during WiringProblem solving to
connect sources to sinks to (manually) canonicalize unused wire sinks to
nodes.  This cleans up a pattern that is necessarily emitted during
Grand Central View Annotation handling where the final sink is created
as a wire (as it does not yet have any driver until the WiringProblem it
creates is solved).  Later, the Grand Central View pass expects to only
see nodes and not wires.  By doing the canonicalization inside the
WiringProblem solver, this avoids needing to rely on a wire2node
canonicalization which may or may not be run.

Put differently, this gets Grand Central views to work with
--disable-opt.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>