[HWMemSim, LowerSeqToSV] Add a flag to annotate array registers with `(*ram_style = "distributed" *)` (#4288)
This adds an option `--add-vivado-ram-address-conflict-synthesis-bug-workaround` to add a workaround that vivado maps BRAM to memory register that has caused an incorrect synthesis result. To prevent vivado from using BRAM for registers, this expliclity specifies the ram style so that vivado selects LUTs. With this PR, registers used as memory are annotated in the following way:

```
  (* ram_style = "distributed" *)
  reg  [15:0] Memory[0:9];
```