[ExportVerilog] Emit VerbatimExpr inline in bind
Change emission of SystemVerilog (SV) verbatim expressions to always
inline into bound instances.  This is specifically done to enable
cross-module references (XMRs), encoded as verbatim expressions, to be
emitted in the bind.  FIRRTL-originating XMRs use verbatim expressions
as they require the ability to use substitutions.

This is part of a broader effort to disallow "upwards XMRs" in FIRRTL.
These are represented using input RefType ports.  These are generally
problematic because an "upwards XMR" may require duplication of the
module in which it is used in order to legalize to an SV XMR.  However,
this means that any preexisting upwards XMR now needs to be converted to
an input port.  By enabling these XMRs to be emitted in the bind, no
modifications are necessary in the module where the bind occurs.
Without this commit, a dead wire would have to be placed unconditionally
in the bind site.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>