[PyCDE] Support for BSP-defined packaging (#4338)
Introduce the notion of a "board support package" (BSP). It controls the top level of the design and assembles the system into a "package". A package generally includes the hardware output (+ any other collateral needed to build the design), the software runtime, and a script (makefile) to build the system and (for some BSPs) run it.