[SV] Add XMRRefOp
Add a new hierarchical reference op, XMRRefOp, that uses a symbol to
refer to a hw::GlobalRefOp that encodes the path to something in the
circuit.  The XMRRefOp includes an optional ability to refer to
something inside a reference via a string.  This is intended to be used
to XMR into external modules.

The XMRRefOp exists to complement the more restrictive XMROp which is a
very direct translation of a Verilog hierarchical reference that only
uses strings.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>