Fix error in reducer test for annotation-remover
In the annotation-remover, the intent was to use grep to ensure that
anotherWire is preseved in the final output, and use filecheck to ensure it's
annotations have been removed. At some point the test was modified and this
intent was lost. This PR corrects this script to it's intended behaviour.