[PyCDE][ESI] Load a cosim connection out of a package dir
For cases where the RTL simulation is running with the package dir as
cwd, provides an easy way to connect to it. Untested since the
integration tests are using the esi-cosim-runner.py tool. Will build
more advanced support for running cosim in the future, after which this
can be tested.