[Handshake] Fix use-after-free bugs in canonicalization patterns.
A couple of the Handshake ops had use-after-free bugs in their
canonicalization patterns. We were calling `eraseOp()` on the root op
within an `updateRootInPlace()` call. This is illegal because
`updateRootInPlace()` must only update the root op in-place, not erase
the op altogether.

These were fixed by moving the `eraseOp()` call after the
`updateRootInPlace()` call.