[CombFolds] Handle array operations in flattenConcat (#4389)
This PR implements a canonicalizer to merge array operations in concat operands. 
In chisel, it's quite common to concatenate vector elements to form integer. With aggregate preservation, 
they are bit-compatible in HW/SV level. For example, 
```scala
circuit Foo:
  module Foo:
    input a: UInt<1>[3]
    output b: UInt<3>
    b <= cat(a[2],cat(a[1], a[0]))
```
is transformed into
```
  assign b = a;
```
with aggregate preservation. Previously this was `assign b = {a[2'h2], a[2'h1], a[2'h0]};`