Handshake builders cleanup and SOST interface
This commit removes a lot of custom builders for Handshake operations,
and re-enables default builders for most operations. The commit also
replaces the sost namespace (along with its operation attributes logic)
with an SOSTInterface that Handshake operations can implement to achieve
the same purpose with less code and more reliability. This in turns
facilitates the deletion of some operation attributes that were
redundant.

A lot of tests still fail on this commit due to largely unupdated
parse and print logic for Handshake operations.