[EmitOMIR] Support fieldID for OMIR targetting fields.
This adds support for OMIR that targets fields. Tracker is updated to
hold a fieldID, which behaves like it does in Annotations: defaults to
0 to mean the whole entity, and greater than zero to index into a
field of the entity. When constructing output OMIR, if the tracker has
a fieldID, the type of the entity being targeted is used to append the
appropritate output field. For vectors, this is an index, and for
bundles, this is a named field. If we don't know how to easily extract
the type of a targeted entity, an error is raised. The common cases of
Wires, Registers, Nodes, and Ports are supported.