[HW] Add FieldIDTypeInterface
Add a rudimentary type interface to the HW Dialect for representing
types which contain aggregates which can be indexed/referred to by a
"field ID".  A field ID is an in-order number of an aggregate type that
includes the top-level aggregate type (field ID zero), all intermediate
aggregate types, and all leave types.  This is an interface that was
created from infrastructure inside the FIRRTL Dialect for its aggregate
types.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>