Verification and type inference for HandshakeOps
This commit adds some verification logic as well as return type
inference for a number of Handshake operations. The latter allows to
shorten the creation of Handshake operations in many places as the
return type(s) need no longer be specified at creation.

The commit also introduces a simple builder for ForkOp and LazyForkOp.