No implicit merge semantics during merge insertion
This commit gets rid of the assumption on backedge ordering during merge
insertion. The MergeOpInfo struct now carries an optional and separate
backedge for merge-like operations that have an index operand (e.g.,
MuxOp). The implemention no longer assumes the index backedge to be
the first operation operand. As a consequence, reconnectMergeOps is
simplified.